Memoir-like class update
Started to make the various bits of content a little more memoir-like, in the sense that they are now formatted by a series of macros which could be redefined in the document itself using renewcommand.

Also added epstopdf to the dependencies, and removed the pdf files for the logos and sigils.

The option titleheight now takes an absolute height, and the default is to set the height to logoboxwidth.